+++ .github/home-manager.nix
@@ -1,0 +1,10 @@
+{ pkgs, ... }:
+
+let
+  repository = /home/runner/work/shadow-nix/shadow-nix;
+in
+{
+  imports = [
+    (repository + "/import/home-manager.nix")
+  ];
+}